cleanup

+++ src/util.h
@@ -1,4 +1,7 @@
-#include "main.h"
+#include <string>
+#include <cstdint>
 
-void TextOut(const int x, const int y, const std::string& text, const uint64_t color = WHITE);
+#include "constants.h"
+
+void TextOut(const int x, const int y, const std::string& text, const uint16_t color = WHITE);
 void PrintStatus(const std::string& status);
